Melhorias e Correções de Bugs

+++ Stock_System/Models/Rolo.cs
@@ -4,7 +4,7 @@ namespace EstoqueRolos.Models
     {
         public string Code { get; set; } = string.Empty;      
         public string Descricao { get; set; } = string.Empty;
-        public decimal Milimetragem { get; set; }
+        public Double Milimetragem { get; set; }
         public int MOQ { get; set; }
         public int Estoque { get; set; }
         public int MetragemWIP { get; set; }
